Update SenseHat_FakeTV.py
The four random effects functions, although each displaying for a random length of time, had a 25% chance of being called.  I wanted to change the probability these four functions can be called.  Lines 48-57 now allow for that with simple error checking. I put emphasis on the 'noise' and 'fader' functions.  At this point, this project is overkill.